requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
PyMySQL==1.1.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
alembic==1.16.2
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.2.1
